+++ frb_rust/src/ffi/mod.rs
@@ -3,11 +3,7 @@ pub type DartAbi = wasm_bindgen::JsValue;
 #[cfg(not(wasm))]
 pub type DartAbi = allo_isolate::ffi::DartCObject;
 
-#[cfg(wasm)]
-pub trait IntoDart {
-    fn into_dart(self) -> DartAbi;
-}
-use std::{ffi::c_void, sync::Arc};
+use std::{ffi::c_void, sync::Arc, ops, mem};
 
 #[cfg(not(wasm))]
 pub use allo_isolate::IntoDart;
@@ -29,6 +25,8 @@ pub type Channel = allo_isolate::Isolate;
 pub mod io;
 #[cfg(not(wasm))]
 pub use io::*;
+
+use crate::DartSafe;
 
 /// see [uuid::Bytes](https://docs.rs/uuid/1.1.2/uuid/type.Bytes.html)
 #[cfg(feature = "uuid")]
@@ -77,4 +75,110 @@ unsafe extern "C" fn drop_arc<T>(ptr: *const c_void) {
     // Dart has ownership of this copy of Arc, and can only share out clones,
     // so this is safe to call exactly once.
     Arc::<T>::decrement_strong_count(ptr as _);
+}
+
+
+/// A wrapper to transfer ownership of T to Dart.
+///
+/// This type is equivalent to an [`Arc<T>`]. The inner pointer may
+/// be None if a nullptr is received from Dart, signifying that this pointer
+/// has been disposed.
+///
+/// Extensions for [`sync::RwLock`] and [`sync::Mutex`] are provided.
+///
+/// ## Naming the inner type
+///
+/// When an `Opaque<T>` is transformed into a Dart type, T's string
+/// representation undergoes some transformations to become a valid Dart type:
+/// - Rust keywords (dyn, 'static, DartSafe, etc.) are automatically removed.
+/// - ASCII alphanumerics are kept, all other characters are ignored.
+///
+/// ## Trait objects
+///
+/// Trait objects may be put behind opaque pointers, but they must implement
+/// [`DartSafe`] to be safely sent to Dart. For example, this declaration can
+/// be used across the FFI border:
+///
+/// ```rust
+/// use flutter_rust_bridge::*;
+/// use std::fmt::Debug;
+/// use std::panic::{UnwindSafe, RefUnwindSafe};
+///
+/// // Rust does not allow multiple non-auto traits in trait objects, so this
+/// // is one workaround.
+/// pub trait DartDebug: DartSafe + Debug {}
+///
+/// impl<T: DartSafe + Debug> DartDebug for T {}
+///
+/// pub struct DebugWrapper(pub Opaque<Box<dyn DartDebug>>);
+///
+/// // creating a DebugWrapper using the opaque_dyn macro
+/// let wrap = DebugWrapper(opaque_dyn!("foobar"));
+/// // it's possible to name it directly
+/// pub struct DebugWrapper2(pub Opaque<Box<dyn Debug + Send + Sync + UnwindSafe + RefUnwindSafe>>);
+/// ```
+#[repr(transparent)]
+#[derive(Debug, Clone)]
+pub struct Opaque<T: ?Sized + DartSafe> {
+    pub(crate) ptr: Arc<T>,
+}
+
+impl<T: ?Sized + DartSafe> ops::Deref for Opaque<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        self.ptr.as_ref()
+    }
+}
+
+impl<T: ?Sized + DartSafe> From<Arc<T>> for Opaque<T> {
+    fn from(ptr: Arc<T>) -> Self {
+        Self { ptr }
+    }
+}
+
+impl<T: DartSafe> Opaque<T> {
+    pub fn new(value: T) -> Self {
+        Self {
+            ptr: Arc::new(value),
+        }
+    }
+}
+
+impl<T: DartSafe> From<Opaque<T>> for DartAbi {
+    fn from(value: Opaque<T>) -> Self {
+        let ptr = Arc::into_raw(value.ptr);
+        let drop = drop_arc::<T> as *const c_void;
+        let share = share_arc::<T> as *const c_void;
+        let size = mem::size_of::<T>();
+
+        vec![
+            ptr.into_dart(),
+            drop.into_dart(),
+            share.into_dart(),
+            size.into_dart(),
+        ]
+        .into_dart()
+    }
+}
+
+/// Macro helper to instantiate an `Opaque<dyn Trait>`, as Rust does not
+/// support custom DSTs on stable.
+///
+/// Example:
+/// ```rust
+/// use std::fmt::Debug;
+/// use flutter_rust_bridge::*;
+///
+/// pub trait MyDebug: DartSafe + Debug {}
+///
+/// impl<T: DartSafe + Debug> MyDebug for T {}
+///
+/// let opaque: Opaque<Box<dyn MyDebug>> = opaque_dyn!("foobar");
+/// ```
+#[macro_export]
+macro_rules! opaque_dyn {
+    ($ex:expr) => {
+        $crate::Opaque::new(::std::boxed::Box::new($ex))
+    };
 }
